drm/v3d: Don't access V3D registers on reset
There is a race between IRQ and reset.  It may happen where the IRQ
handler is called during the reset due to a job timeout.  So, in the IRQ
handler, holding the reset lock during the access to the V3D registers
is needed to avoid reading meaningless values from V3D registers.

Signed-off-by: Yukimasa Sugizaki <ysugi@idein.jp>